backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
pydantic==1.10.9
alembic==1.11.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
cryptography==41.0.3
celery==5.3.1
redis==4.5.5
requests==2.31.0
prometheus-client==0.17.0
httpx==0.24.0
